requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-bulk-admin==0.1.1
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
idna==3.4
kavenegar==1.1.2
Pillow==9.5.0
pyasn1==0.5.0
PyJWT==2.7.0
python-jose==3.3.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.2
